Step 5d follow-up: indexer_pool kernel_page_size + remove double-divide
Two fixes on top of the initial step-5d commit:

1. DeepSeekV4IndexerPool: add `kernel_page_size` param (defaults to its
   own page_size for CUDA compat). NPU buffers now allocate with
   shape (npu_num_pages, kernel_page_size=256, 1, index_head_dim)
   instead of (..., c4_page_size=64, ...). The previous shape made
   cmp_kv.shape[1] = 64 != ori_kv.shape[1] = 256 and triggered
   `aclnn 507015 / DDR address out of range / aicore exception` on
   the FIRST npu_quant_lightning_indexer call.

2. _forward_compressed (attention): drop the `// page_ratio` divide
   on `cmp_block_table` when it comes from the step-5c slab. The
   divide was a leftover from step 5b (when cmp_block_table came from
   raw kv pool page indices in [0, 2349)); slab values are already
   in kernel-view page space [0, 587) so the divide silently shifted
   the kernel's read to the wrong page (e.g. slab page 146 read at
   page 36 = stale/empty kv → zero c4 contribution).

217 verification with SGLANG_DSV4_NPU_REAL_COMPRESSOR=1 +
SGLANG_DSV4_NPU_SPARSE_ATTN=1 + SGLANG_DSV4_NPU_REAL_INDEXER=1:
- ASCEND_LAUNCH_BLOCKING=1, 2-token "Four " probe -> HTTP 200,
  output [19, 0] in 6.9s. Kernel chain runs end to end: compressor
  writes c4_indexer_kv_pool int8+scale via npu_scatter_nd_update_,
  npu_quant_lightning_indexer returns topk, _forward_compressed
  feeds it into npu_sparse_attn_sharedkv. Zero `aclnn / aicore
  exception` events.
- Async (no LAUNCH_BLOCKING): server hangs on prefill (forward never
  completes, curl 300s/600s both disconnect; no kernel errors). The
  sync sentinel kwargs are identical to the working sync run, so
  it's an event-ordering issue (likely npu_scatter_nd_update_ async
  write not visible to the subsequent kernel read without an
  explicit event wait). Step 5e territory.

Indexer-side dbg dump (gated by SGLANG_DSV4_NPU_SPARSE_ATTN_DEBUG=1)
added to _forward_npu_fused — prints q/k/k_scale/weights shapes +
dtypes, block_table value range, actual_seq_lengths, index_topk
before every kernel call. Identical to what iforgetmyname feeds
forward_npu_dsv4_fusion.

+++ python/sglang/srt/hardware_backend/npu/attention/deepseek_v4_ascend_backend.py
@@ -549,12 +549,14 @@ class DeepseekV4AscendAttnBackend(
         # Reshape cmp_kv to share page_size with ori_kv before the kernel call.
         # main's V4 pool layout: c{N}_kv_pool buffer is (num_pages, page_size//
         # ratio, 1, dim) so each native page holds page_size//ratio compressed
-        # tokens. iforgetmyname / the aclnn kernel expect cmp_kv to share its
-        # page_size with ori_kv (=global page_size). We slice the buffer to a
-        # ratio-aligned native-page count and view it as (N_kernel,
-        # global_page_size, 1, dim); the corresponding cmp_block_table is
-        # `block_tables // ratio` (one kernel page covers `ratio` global raw
-        # pages worth of compressed slots).
+        # tokens. The aclnn kernel expects cmp_kv to share its page_size with
+        # ori_kv (=global page_size). We slice the buffer to a ratio-aligned
+        # native-page count and view it as (N_kernel, global_page_size, 1, dim).
+        #
+        # cmp_block_table values: step 5c slab (`req_to_token_c{N}_pages`)
+        # already gives kernel-view page indices in [0, N_kernel), so no
+        # further `// page_ratio` divide is needed — the divide was a leftover
+        # from step 5b when block_table came from raw kv pool page indices.
         ori_page_size = ori_kv.shape[1]
         cmp_native_page_size = cmp_kv.shape[1]
         cmp_block_table = getattr(
@@ -571,7 +573,8 @@ class DeepseekV4AscendAttnBackend(
             cmp_kv = cmp_kv[: n_kernel * page_ratio].reshape(
                 n_kernel, ori_page_size, *cmp_kv.shape[2:]
             )
-            cmp_block_table = (cmp_block_table // page_ratio).to(torch.int32)
+            # Slab already in kernel-view page space — no divide.
+            cmp_block_table = cmp_block_table.to(torch.int32)
 
         attn_kwargs = dict(
             cu_seqlens_q=fm.actual_seq_lengths_q_pa,
